Prevent debug log when hiding cart combine
Addresses the issue described at:
https://www.zen-cart.com/showthread.php?228272-Help-with-Error-Log-(PHP-Notice-Undefined-variable-zc_check_basket_before)

This takes the shortest/simplest action to correct the issue by use of logic
instead of additional code.  Any other change simply adds more code to file
but may be desired to ensure updaters take at least one action which would
prevent this issue in a common otherwise unmodified store.

This is applicable to V1.5.7 and previous version as well, but likely
is not made known unless using strict PHP operation as it does not appear
to have been revealed until a high enough version of php was used and
someone reported it.