Pre-assign value
This at least prevents the message that the variable is not defined which
was identifeid at: https://www.zen-cart.com/showthread.php?228272-Help-with-Error-Log-(PHP-Notice-Undefined-variable-zc_check_basket_before)

This is additional action that isn't *required* to be present based on the
change of the previous commit.

It would resolve the issue if incorporated alone.