Add a value check to resequence of configuration value
This adds a check against `!empty` to ensure that `$zc_check_basket_before`
has been defined even though its value is expected to represent the quantity
of product in the cart which is expected to be zero or greater. Therefore
it does not account for a quantity greater than 0 as well. Such a consideration
may be necessary for a site where returns are provided at checkout and the
quantity of returns is for some reason expected to be a negative number.