Allow other multiple addresses to process
If there is more than one recipient for an email, but only one of the recipients requested not to receive an email, existing code would stop processing emails beyond that individual. Worst case would be that the first of many did not want an email and therefore none would receive one.  This change continues to interupt the process of sending the email; however, will allow the code beyond the loop to process error information and to use the notifier that is present (which may have an effect on "output").