Begin numbering so that all items on first page area always selectable.
Through observation this counter needs to begin at a value of 0 so that
when a quantity spans more than one page, that the last item on the first
page can be selected allowing the page to reload with it selected and
editable. There may be other arrangements, but this fixes the issue with
the code continuing to exist here.

Fixes #4489 for ZC 1.5.7

Align to execute inner code when "page" not provided or on first page

This aligns the code section to execute the internal code when either the
particular admin page is loaded with a specific item to be displayed with
the page parameter not identified or set to a value of 1. What this does is
allow the appropriate page to be displayed based on the associated settings
when only the item's identifier is known.