Begin numbering so that all items on first page area always selectable.
Through observation this counter needs to begin at a value of 0 so that
when a quantity spans more than one page, that the last item on the first
page can be selected allowing the page to reload with it selected and
editable. There may be other arrangements, but this fixes the issue with
the code continuing to exist here.

Fixed #4489