Add new address formats and update old ones.
Address formats 1 and 3 removed as all countries not have $country on a
separate line.

Comma removed from format 2 as all addresses I have found do not appear
to have a comma. This is then the same as format 7 so countries using 7
will be moved to 2.

Format 1 replaced with Spain including brackets for state
Format 3 new format $postcode $city $state
Format 7 replaced with Hungary $city $street / $postcode
Formats 8-15 new formats.

Note if countries with no postcode are to be displayed with no postcode
An additional 2 postcode will be required duplication 4 and 6 with the
$postcode omitted. Currently will display the content of postcode at the
end of a line (format 4) of on a separate like (format 6)