square SCA compliant
Copy of Rocketsites changes to Square modules to make them SCA
compliant.
as per request The-end-of-Zen-Cart-in-Europe-SCA-kills-payments from
@swguy

I have copied the changes to the 1.5.8 stream. I think I have formatted
it correctly, please check.
Do you want additional comments in the code as I have removed the //SCA
changes 1 of 3 etc. as i thought unnecessary as git shows you the
changes