Observe the name change for Türkiye
United Nations as of 2 June 2022 recognizes the name change and it should therefore be adopted here.

Keeping in line with the way that countries are now handled in Zen Cart, a
new install will have Türkiye at country 215, while an upgrade will update
the country name if it's countries_iso_code_3 is present. This is in line
with the explanation/direction provided in issue #3664.

If the country has been deleted, then reinsertion would need to be by some
other path.

Previous attempt to save the mysql_utf8.php file encoded in UTF-8, which
did not resolve properly. Force pushing the change with the file encoded
to ANSI before pasting the content successfully displayed in the update
file.