Check $gInfo and gID set on insert
fix 5144
Modify insert button to only output the gID if $gInfo is set.
Modify the new form to only output the gId if $_GET['gID'] is set.
Have done it this way so if you hit cancel in the form you are returned
to the group you were editing. Not the first group of the page you are on.