Correct PayPal Checkout Flow
Changes to set the $_SESSION\['cartID'] if the shopping cart has not had
items added to it.

Alter logic to always find the Cart content count before restorecart
method. os that if the cart is combined  but no message is to be
displayed The before count is available for subsequent logic.