Fix PayPal require USD
This would be better written as a function. Having said that where do I put it so that it can be called by all these routines.

I have checked to see is 'USD' is a valid currency and if not then fetched the conversion value using the functions_exchange_rates. If that fails to find a rate it will create a warning file with $rate not found similar to now. It will still process the transaction as it does now.

Reworked as function

Added paypaldp

lat9 feedback