v158: Correcting `zen_get_products_description`
As identified in [this](https://www.zen-cart.com/showthread.php?229078-Product-Description-Does-not-Show-up-in-All-Products-and-other-Pages) Zen Cart forum posting, the description returned by the subject function is always an empty string.  That's because the fall-back (when no language is set) was setting a variable named `$language` instead of `$language_id`, so the $language_id used for the query was always 0.