Include Plugins in Search?
Incorporates the upgraded changes made in #4486.

With the addition of Plugins to the admin side of the store, searching
for information within the plugins remains advantageous. However, with the
addition of plugin versions, all plugin software is searched when looking
at the admin and not just the latest or active version nor is there a way
to exclude searching the plugins on a per search basis.

This adds a checkbox to add the zc_plugins directory into the applicable
search which currently is all files, all admin and just the plugins
searches plugins directory. An alternative solution that was temporarily
locally implemented was to add options into the dropdown to describe
performing the search without plugins; however, the list becomes that much
longer and more code to be added.

The choice of add plugins was made to demonstrate that if plugins
was already chosen, it would automatically and already be included, but
if plugins were not originally part of the associated directory (admin for
example doesn't have a plugins sub-directory) then it incorporates the
plugins into the search.