Add inverse selection feature
There are times when there are just a few files to keep, but many others to
be removed. Selecting a few to then inverse the selection and then choose
to delete all selected improves efficiency.

Added a button to inverse the selection(s). Also updated the script within
the main file to not be encapsulated around HTML comment blocks.

Updated as version 3.0.2. Tested the upgrade process which successfully
completed.