Ignore products in subcategories
Allow categories count to ignore products in sub categories.
Added flag COUNT_SUB_CATEGORIES to allow category tree to only count the products in the category specified.
Modified Category tree to check it the new flag exists and if not continue to count the products in subcategories.
Modified functions_categories to add new parameter to zen_count_products_in_categories to allow subcategories to be ingnored in the count.