"Disable" ask-a-question if disabled for product's type
If a site has disabled the `ask_a_question` link (via Catalog :: Product Types :: Layout Settings for a product type), the page still renders.

It's likely that the site's disabled the link because they don't want to deal with additional emails coming from that page, but a knowledgeable person could still navigate to that product-specific page via their browser's address bar.

This update simply redirects back to the product's `_info` page if the feature is disabled for its product-type.  There's also a change to the redirect if a product is not found, since the `_info` pages expect to see a `products_id` parameter, not a `pid` one.