Include Laravel in Search?
With the addition of Laravel to the store, searching for information within the Laravel code is advantageous. However, the laravel directory (integral to many operations and to become even more so) has been excluded from search.

This adds a checkbox to add the laravel directory into the applicable search which currently is all files and also adds a search option to search just the laravel directory (to potentially exclude other code that uses code from that directory).

The choice of add laravel was made to demonstrate that if laravel was already chosen, it would automatically and already be included, but if laravel was not originally part of the associated directory search then it incorporates the laravel directory into the search.

Fixes #5513. May have conflicts with 5512 because they were added independently. Intention was for this (Laravel) to "follow" any plugin related content at the applicable location. I can rebase and force push if necessary to eliminate the conflict if desired/requested.