Stop duplication of defines if Square Webpay is isntalled.
Modified to stop the already defined messages appearing if you have installed Square webPay. It defines a number of variables a constants that are already defined in this language file.

As I assume that square will be removed completely from zen cart at some point I thought it better to modify this file rather than have code in a new webPay module .