Add a notifier for `SHOW_` . `$info_page` . `_ASK_A_QUESTION`
A unilateral decision was made that a site owner must modify core/template
code in order to allow being asked a question about a specific product
just because they have turned off the option to display on the associated
product page. This adds a notifier to allow being asked a question about
a product as determined permissable in the observer.

Defaults to continue as last determined to be handled, but allows
overriding that. At least one situation where this feature is desired
is in turning off the option so it doesn't display on the product type,
but including a link in email to click to ask a question about the product.

Without this change, then some other code edit must be done on a more
permanent basis instead of just adding an auto loading observer.
page