Restore cart identification in whos online
Decoding session data with `base64_decode` appears to no longer return all
data that was initially encoded. As a result, the search of the decoded
session data could likely fail finding one or more product in the list.
This takes advantage of the added class method `inspectSessionCart` that
temporarily becomes the session to collect the data and return the session
information back.

Passing an empty value for the `session_id` because the code already requires
the `session_data` to be non-`empty` and the method is written to operate so
long as either is non-`empty`.

Added a variable (`$inactive`) to support a single calculation instead of
the same evaluation multiple times.

Simplified logic as either there is a return because there are no rows in
the cart or the return is because of the activity status.

Fixes:
https://www.zen-cart.com/showthread.php?229530-Who-s-Online-not-showing-active-carts&p=1395104#post1395104