v158 ajax closure and refactor
Call the `includes/application_bottom.php` file applicable to the side of
the store used to open the ajax load.

Do away with nested if/elseif logic where return is used for true evaluation.

Eliminate single line if statements and add parentheses to support
expanded logic grouping.

Eliminate the parentheses around the filename for calls of `require`.