Resolve #5799
So since the determination was to not undo the previous entry, this instead
extends it.

Offers a new variable that can be overridden in advance of the associated
code section allowing execution of the inner code section but not
requiring a `$products_image` to hold a value at the point where this
begins. As a result of work that has historically gone into offering
the code the way it was, this extends that previous capability that has
been removed by #5656.

Fixes #5799