Fix layout of search_box elements.
The container, outside of search_box.php for example customers.php, is
col-md-4 which means it will be 4/12 width in medium breakpoint, and full
page width in smaller breakpoints.  The columns here were also set to
col-md-4, so they were 4/12 each in a space that was already quite narrow.
In small breakpoints they jumped to 100% width which is also not optimal.

The approach here is:
- extra small and small, 50% width so the checkboxes lay out side by side.
- other breakpoints - 100% width so the checkboxes stack on top of each
other in the narrow screen space they have been allotted by their parent.