v158b Add string PRODUCT_URL_REGEX sanitizing
Original design of the admin sanitizer associated with PRODUCT_URL_REGEX was
written to directly support only the data type that is specifically used by
Zen Cart which is a language based array of values. Use of this filter type
on a non-array variable causes errors.

Three methods of resolution came to mind:
1. Other software attempting to use this filter should make the data an
array, filter it, then return it to not being an array.
2. Add/expand filtering by use of class overriding.
3. Add an additional filter that only addresses stringss (not arrays). This
appears would complicate the code and control. Either the new code would
reference the existing (logging a "non-existent" sanitization) or would
effectively duplicate the code such that a change in one would require a
change in the other.
4. Expand the method's code to perform the same operation but with the
data formatted solely as a string instead of an array and without
recursion.

As part of the feedback from the initial poster, also corrected the
capitalization of the call to the method `setCurrentArrayName` to
ensure systems that have begun enforcing or will enforce this.

I normally would have placed the smaller/shorter code portion first and
then allowed the `foreach` loop to follow; however, based on so many
micro-optimisations recently implemented I am not sure what
(re-)sequencing is "best". Considering the base code was expected to
process an array, I put/kept that first so that inner code would not
always get bypassed during normal operation.

I also left the assignment to $urlRegex before the early escape to align
with perhaps old coding style recommendations of defining a variable as
soon as possible within the code. A micro-optimisation would be to
move that variable assignment to after the first check so that it is
created only when it would become useful. Though perhaps there are other
reasons to declare it each method call.

Fixes #5849