v158b Refactor Admin Sanitizer and add String Only sanitization
In association with fixing #5849, applied the same type of correction
to other unique filters that only managed array type sanitization.

With incorporation of a change, took an opportunity to refactor the
code using some of the principles that have been applied by others in the
community.
early escape
remove empty lines
use elseif instead of else if