category_product_listing: valid HTML and notifiers
If a site has set `SHOW_COUNTS_ADMIN` to 'false', a categories/products listing that is a mixture of categories generates invalid HTML due to the missing table-data column for the Quantities column.  For example, with the demo products, search for `DVD` to bring up such a mixture.

In addition, while there are notifications to allow the gathering of additional category-related fields during an insert/edit, there are currently no notifications to allow additional columns to be added to a category-listing in the admin.

This PR defines those notifications.