Correct "Calculate Tax" when "None" (#5878)
The `include_tax` property is set to the value of the configuration setting, a 'true' or 'false' strings, so its value is always interpreted as _truthy_.

Add a strict comparison to determine whether/not the taxes are to be included.  As the module currently sites, tax is _always_ included when the "Calculate Tax" setting is "None".