Add ability to count distinct products (#5910)
Currently the category count counts the number of products in each sub category individually. This means that if you have a product in more than one sub category of the parent category it is counted once for each category that it is in. If like me you have products that you put into many sub categories to allow searching by a feature of the product the category count can be inflated.
This mod counts the distinct products_id(s) in a category and all its sub categories.