Capture payment module final status in order (#6000)
If a payment module, e.g. `authorizenet_aim` sets a different status during its `before_process` method, that status is not initially recorded into the order -- which can cause customer confusion.

For instance, if an a.net/aim transaction is being held for review, a specific a.net/aim order-status can be set.  Currently, the order's initial status (viewable by the customer) indicates that the order is awaiting processing, but a subsequent status-update by a.net/aim sets the order's status to its "Under Review" setting.

This update corrects that issue.