Protect admin against invalid cPath (#6024)
If an invalid `cPath` is provided to the category_product_listing (e.g. cPath=999999), the listing page displays an empty category to which other categories or products can be added.

If, in this case, a category is added that newly-added category's `parent_id` reflects that invalid (non-existent) category; if a product is added, the product's `master_categories_id` reflects that invalid category.