Adding notifications to the `order_total` class
Sometimes, a payment-module needs insight into the details of an order to properly form its request to the associated payment processor.

This PR adds notifications to the two order-total building methods in the subject class to enable an observer to gather the changes made by each order-total module present in the order to make that determination.