Add option to copy specials to copied product.
Visibility of Copy To panels depend on radio buttons.

When the Link or Duplicate radio buttons are changed, show/hide
the panels with link-only and duplicate-only classes by
adding/removing the hiddenField class.
Would like to use an expando attribute like data-link-only="true"
instead of non-functional classes link-only/duplicate-only, but
infoBox does not allow this.