Admin: Additional BS4/BS5 CSS classes in stylesheet.css
Include some basic Bootstrap 4/5 CSS Utility classes in the admin's base `stylesheet.css` as an aid in transitioning admin tools to a more recent version of Bootstrap.  The currently-used classes are also removed from other admin .css and .php files.

The classes are in alphabetical order towards the end of that file.

Note: Also removed duplication of `.clearBoth` definition.