Retire use of `banners_on_ssl` field
Now that SSL is a given, and switching between https-and-non-https is no longer common, retiring the use of this flag seems prudent.
Leaving the field in the database (for now) to avoid backward-compatibility breaks such as old templates/modules making queries to that field. Its default is 1 (on), which is sensible.

Ref discussion #6627