Add an order-class notification for address overrides
Add a notification to enable an observer to override customer/delivery/billing addresses for an order created from the 'cart' in preparation to be recorded in the database.

The addition of this notification will allow the next-generation of "Edit Orders" to make use of the base order-class processing in the gathering of the information for the to-be-updated order.

Also includes some minor reformatting, breaking out the `country` elements into one-line for each array key.