Encap plugin installer logic error?
@zcwilt

It would seem that it checks here for `$versionDir/installer/Installer.php` and then tries to load `$versionDir/Installer.php`
Plus the use of `new Installer()` missing the `\` as in `new \Installer()` potentially clashes with the main `Installer` class, which has none of the methods on it that are expected by the code using this Factory to generate an "installer" object for.

So, I'm not convinced that this PR is even a proper "fix", but it does point out where there are mismatches that need attention.

What was the "Installer/Installer.php" fallback intended to be used for?  Is there a use-case for other than `BasePluginInstaller`, which `ScriptedPluginInstaller` uses?