Allow PSR4 naming for auto-instantiated Observer class filenames
# Auto-Instantiation of Observer Classes
At Zen Cart's auto-loader load-point `#175` we look for observer files whose filename begins with `auto.` and then `include()` the file, and then `instantiate` the class into a global object of the name of the class.

## Current:
Note the `auto.` prefix on the filename, and the `zcObserver` prefix inside the class as the class name.
| File name | Class name | Global object var |
|----|----|----|
| auto.ProductMusicObserver.php | zcObserverProductMusicObserver | zcObserverProductMusicObserver  |
| auto.MapMinAdvPriceAdmin.php | zcObserverMapMinAdvPriceAdmin | zcObserverMapMinAdvPriceAdmin |
| auto.downloads_via_aws.php | zcObserverDownloadsViaAws | zcObserverDownloadsViaAws |
| auto.non_captcha_observer.php | zcObserverNonCaptchaObserver | zcObserverNonCaptchaObserver |

## New
**(These are in addition to the above-supported original conventions)**

Essentially the change here is to 1) change `auto.` to `Auto` and 2) have the class name match the filename.
Note that the global object var still adds the `zcObserver` prefix as a distinguisher.
| File name | Class name | Global object var |
|----|----|----|
| AutoProductMusicObserver.php | AutoProductMusicObserver | zcObserverAutoProductMusicObserver |
| AutoObserverTagManager.php | AutoObserverTagManager | zcObserverAutoObserverTagManager |

### added convenience helper
Let's say you have a class that started out as NOT auto-instantiating, but later wanted to MAKE it auto-instantiating...
Previously you had to both add the `auto.` prefix to the filename and the `zcObserver` prefix inside the file.

Now you can simply add the `Auto` prefix to the filename, and it will look for the class name both with and without the prefix.
For example, these would all work:
| | File name | Class name | Global object var |
|----|----|----|----|
| legacy | auto.ObserverTagManager.php | zcObserverObserverTagManager | zcObserverObserverTagManager |
| new | AutoObserverTagManager.php | **AutoObserverTagManager** | zcObserverAutoObserverTagManager |
| convenience | AutoObserverTagManager.php | **ObserverTagManager** | zcObserverAutoObserverTagManager |