Changing admin image display to size via CSS instead of HTML attribute
This change allows the image to be resized dynamically via CSS instead of hardcoded via HTML <img> attributes.

For example: the pictures in the category listing would be stretched to fit both the height and width of the image settings as designated in the admin backend. This change moves the setting of the image's height and width to CSS where appropriate (ie. where the image's height or width are set in the function call).