Change hashing functions to use modern methods
`md5()` -> `hash('md5')`
Not currently using `sha1()` directly: we're already using `hash('sha1')`.

The bulk of our use of `md5()` is just to prepare a token, or to identify one element from another with some uniqueness.
So this is not really a change that affects cryptography per se.
It merely steps away from using the (potentially-being-retired) md5() function.