Adds hint about the MFA to the confirmation screen
To allow people who may not have the ability to scan the QR code but who can copy/paste the code into an MFA tool, this will show the generated MFA secret to copy/paste the secret into their designated service. This prompt will ONLY appear during the initial setup of the MFA. It will not appear again.

Signed-off-by: Paul Williams <retched@hotmail.com>