Change description of MFA to show SITE_NAME
Since a person can have multiple services installed on their domain which may use MFA, this edit changes the description of the MFA Handle to better represent what it entails.

In this case, it is the store's name (retrieved from the DB Constant `STORE_NAME`) and the name of the software (ZenCart). For example: if the store's name is "Blockbuster Video", when you scan the MFA QR Code, it will show as "Blockbuster Video - ZenCart" in the MFA tool. This change does NOT affect any already issued MFA tokens as it is only changing the Issuer field and does not touch anything of the issued Secret.