Changing admin image display to size via CSS instead of HTML attribute for products (#6866)
* Changing admin image display to size via CSS instead of HTML attribute for products

In my other edit in #6793, I made this change to increase the size of the images while keeping their proportions. I failed to realize I had only changed the category icon in that edit. This change also changes the images of the product listing rows.

* Changing hardcoded object-fit to Bootstrap like

This change adds the Bootstrap `object-fit` classes to the admin area stylesheet to be invoked by the product and category images.

I changed the category header icon and category product listing images to the same ones.