Send password-reset URL instead of new password
When a customer requests a password reset, the old behavior was to generate a new password for them, reset it to that password, and email that password to the customer. These emails often got lost in spam traps, and if the email never arrived then the customer could never login again even if they remembered their old password because their old password had been replaced.

This updated logic now generates a temporary reset token and emails that token in the form of a URL to click to take them to the password_reset page where they can choose their own new password. Once the token is validated and the user selects a (suitable-length) password, their account is updated to use that new password, and the token is deleted.

Tokens also auto-expire after a configurable timeframe, so that old links in old emails don't present a security risk.

NOTE: Retires notifier `NOTIFY_PASSWORD_FORGOTTEN_CHANGED` which is superceded by `NOTIFY_PASSWORD_RESET_URL_SENT`
But the same former notifier hooks used with the account_password page (when user chooses their own new password) still fire, for the benefit of 3rd party integrations.