Refactor categories_ul_generator class to allow specifying max_level
One can now override any of the configurable properties of the class without extending it.
And one can now specify a $max_levels value when calling `->buildTree()` to limit the levels to 1 or some other value, instead of getting the entire nested list.
Also caches the db query for 5 minutes, to improve performance when site is getting higher traffic volume.