Add cron for activating enabled products This cron specifically activates products as-of the current minute (actually to the "second") when the cron job runs, by passing `false` to new `$useMidnight` parameter.
If one wanted to bypass using the cron job but set activations to happen "at the current minute", passing `false` to the new `$useMidnight` parameter where this function is called in `/includes/init_includes/init_special_funcs.php` would accomplish the same.

These changes are relatively straightforward because the `products_date_available` field is a `datetime` field already, supporting not just a date but also time.
So, as long as that field is set to the desired minute (or second) for an activation time, the next time this function runs with `$useMidnight` set to `false`, the product will be activated.
(Previously it would activate "all products whose date matches today's date, regardless of minutes/seconds", meaning "even products which activate 2 hours from now would be activated", which might be undesired in some cases.)

Closes #7027 discussion/request.