The actual correction
Besides a teeny bit of reformat/refactor ...

Moving the order-address overrides into an in-module observer, watching for the order's address-creation overrides.  That processing, which used to be at the beginning of `ec_step2_finish` is now present in `notify_order_cart_address_overrides`.

An additional protected class property (country_state) is created so that the override processing has a place to store its country_id and state_id for follow-on use by `ec_step2_finish`.