Invoice/packingslip/confirmation email: Don't include empty telephone number
If a site doesn't require a telephone number (i.e. Configuration :: Minimum Values :: Telephone Number is set to 0), empty telephone numbers are still displayed on an order's invoice or packingslip and are also included in the order's confirmation email.

This PR includes the telephone number in those 3 spots only if the value isn't "empty".