Upload: Lowercase file-extension only
The filename sanitization for PR #6897 was a bit overreaching in its lower-casing of the entire filename, as the goal was to avoid issues with JPG vs jpg files.

This has caused problems in the DbIo plugin (https://github.com/lat9/dbio/issues/240) and possibly others that upload files as part of their processing.

This change updates the sanitization so that only the uploaded file's _extension_ is lowercased.