Add zen_get_buyable_product_type_handlers() function
Retrieves an array of product page names that identify buyable products.
This allows us to mark a page as containing a product which can be allowed to add-to-cart or buy-now with various modules.

eg: `product_info`, `product_free_shipping_info`, etc.
This avoids hard-coding a list of product-types, and supports dynamic detection of custom product types.

Use-Case: Some payment modules (and other services) want to know what kind of page the shopper is on. Being able to indicate that the current page is one from which add-to-cart or buy-now should be offered, allows that ability.