Correcting out-of-sync `product_id` vs. `self::$data`
Something (and I can't explain it) is causing the `self::$data` element to get out-of-sync with `$this->product_id`.  Correcting the product_info page display by ensuring that the two values line up when returning the `getID()` value.

Note, too, that calls to either `withDefaultLanguage` or `forLanguage` had the side-effect of removing the 'lang' array from `self::$data`.  That's been correcting by moving the `unset` of that field from the common `getDataForLanguage` method into the above two methods so that the removal affects only the public `fields` array.