Further corrections
When a left-sidebox is in use, in my case "Featured", it's possible that the global `$product_info` object that the (er) "product_info" page is checking contains the `product_id` of the current product, but the associated static `$data` is no longer associated with that `product_id`.

This update changes the class' `$product_id` property to also be static, so that it stays 'in sync' with its `$data`.