Change to see if first comment is not empty
If the customer or admin checks out without placing an order note, it still creates an empty row in Order Status History before any other module adds their note from payment. So, this will only check the first line that matches and print that in the area. (This way, you avoid showing payment module notes in the admin area in the orders sidebar.)